XORdecrypt
